+++ src/main/java/seedu/duke/DukeException.java
@@ -1,5 +1,8 @@
-package Duke;
+package seedu.duke;
 
+/**
+ * An exception resulting from invalid inputs
+ */
 public class DukeException extends Exception {
     public DukeException(String errorMessage) {
         super(errorMessage);
