+++ src/main/java/seedu/duke/ToDo.java
@@ -1,4 +1,4 @@
-package Duke;
+package seedu.duke;
 
 import java.util.HashMap;
